+++ chatbot/utilities/quine.py
@@ -82,15 +82,13 @@ class Quine:
 
 
 if __name__ == "__main__":
-    base_directory_in = r"C:\Users\jonma\github_repos\jonmatthis\chatbot\chatbot\assistants\introduction_assistant"
+
+    base_directory_in = r"C:\Users\jonma\github_repos\jonmatthis\chatbot\chatbot\discord_bot"
     quine = Quine(
         base_directory=base_directory_in,
-
-        # exclude irrelevant code so you can fit results in standard context size
         excluded_directories=["__pycache__",
                               ".git",
-                              "utilities",
+                              "output"
                               ],
-
         included_extensions=[".py"])
     quine.generate_quine()
